+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 5. `RateLimitProvider::isLimitReached()` was removed in favour of `RateLimitProvider::rateLimitTimeout()`.
     A new method has to return a number of seconds to wait while rate limit is held.
 6. `Dsl\MyTarget\Domain\V1\Statistic\DatedStat` was completely removed as it was redundant in the hierarchy.
+7. `BannerOperator::findAll()` signature was changed.
 
 #### Fixes